+++ src/FilterSubscription.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace iTRON\wpHooksDispatcher;
+
+interface FilterSubscription
+{
+    public function unsubscribe(): void;
+}